feat(astro-kbve): scoped alerts in Argo resource-detail Grafana tab
Follow-up to #10659 (alerts island on /dashboard/grafana). Filters the
global alerts list down to whatever resource is open in the Argo drawer
so the Grafana tab shows only alerts that mention that pod (or its
namespace, for non-Pod kinds).

Why: the cluster-wide alerts panel is great for an overview but useless
when you've already drilled into a specific pod and want to see whether
it's the one firing. The drawer's other tabs (Events, Diff, Logs, Pod
metrics) all narrow to the selected resource — alerts now match that
contract.

Layout:

  apps/kbve/astro-kbve/src/components/dashboard/grafanaAlertHelpers.tsx (new)
    - Extracts severityColor / relativeTime / summarizeLabels / AlertRow
      out of ReactGrafanaAlerts.tsx so both the cluster panel and the
      Argo drawer panel render the same row component.
    - New filter helpers: alertMatchesPod(a, ns, pod) and
      alertMatchesNamespace(a, ns). Pod match falls through to namespace
      when the alert has no pod label, so cluster-level rules that fire
      on a namespace as a whole still surface in the right pod's drawer.
    - Recognizes both `namespace`/`kubernetes_namespace` and
      `pod`/`kubernetes_pod_name` label spellings — different exporters
      use different conventions.

  apps/kbve/astro-kbve/src/components/dashboard/ReactGrafanaAlerts.tsx
    - Drops the local AlertRow + helper duplicates; imports them from
      the shared module. No behavior change on /dashboard/grafana.

  apps/kbve/astro-kbve/src/components/dashboard/ReactArgoGrafanaPanel.tsx
    - New `scopedAlerts` state, second useEffect that calls fetchAlerts
      (60s cache shared with the main page) and filters by the resource
      selector.
    - Compact "Alerts (N)" block rendered above the per-pod stat cards
      (Pod kinds) or above the cluster-overview-link hint (non-Pod
      kinds). Uses the same AlertRow component for visual consistency
      with the main page.

Load discipline:
- fetchAlerts hits the same per-user 60s localStorage cache the alerts
  island already populates, so opening a drawer right after the main
  /dashboard/grafana page renders is a free read.
- Filtering is client-side — no extra Grafana proxy hit per pod.
- Empty result (no alerts mention this resource) renders nothing —
  drawer stays compact.
